将合约份数改为小数

+++ src/main/java/com/fota/trade/domain/ContractCategoryDO.java
@@ -16,8 +16,8 @@ public class ContractCategoryDO {
     private String contractName;
     private Integer assetId;
     private String assetName;
-    private Long totalAmount;
-    private Long unfilledAmount;
+    private BigDecimal totalAmount;
+    private BigDecimal unfilledAmount;
     private Date deliveryDate;
     private Integer status;
     private Integer contractType;
